Day 8

+++ database/migrations/2024_07_09_012647_create_amnhacs_table.php
@@ -11,19 +11,21 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('amnhacs', function (Blueprint $table) {
-            $table->increments('id');
-            $table->string('ten', 255);
-            // Foreign keys
-            // tạo trường dữ liệu
-            $table->integer('id_nhacsi')->unsigned()->nullable();
-            // tạo foreign keys
-            $table->foreign('id_nhacsi')->references('id')->on('nhacsis')
-                //->onDelete('cascade')
-                ->onDelete('set null');
-            $table->text('mota', 255);
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('amnhacs')) {
+            Schema::create('amnhacs', function (Blueprint $table) {
+                $table->increments('id');
+                $table->string('ten', 255);
+                // Foreign keys
+                // tạo trường dữ liệu
+                $table->integer('id_nhacsi')->unsigned()->nullable();
+                // tạo foreign keys
+                $table->foreign('id_nhacsi')->references('id')->on('nhacsis')
+                    //->onDelete('cascade')
+                    ->onDelete('set null');
+                $table->text('mota', 255);
+                $table->timestamps();
+            });
+        }
     }
 
     /**
